just productDB to go
fixed several typos breaking most of the tests, still need to fix productDB

+++ ado_lab_2023-stewartl71-master/ado_lab_2022-stewartl71-master/MMABooksADO2022/MMABooksBusinessClasses/Product.cs
@@ -12,7 +12,7 @@ namespace MMABooksBusinessClasses
         //constructors
         public Product() { }
 
-        public Product(string id, string description, Nullable<double> unitPrice, Nullable<int> onHandQuantity)
+        public Product(string id, string description, Nullable<decimal> unitPrice, Nullable<int> onHandQuantity)
         {
             ProductCode = id;
             Description = description;
@@ -22,7 +22,7 @@ namespace MMABooksBusinessClasses
         //instance variables
         private string productCode;
         private string description;
-        private Nullable<double> unitPrice;
+        private Nullable<decimal> unitPrice;
         private Nullable<int> onHandQuantity;
 
         //get/set
@@ -53,7 +53,7 @@ namespace MMABooksBusinessClasses
                     throw new ArgumentOutOfRangeException("description must be between 1 and 50 characters");
             }
         }
-        public Nullable<double> UnitPrice {
+        public Nullable<decimal> UnitPrice {
             get
             {
                 return unitPrice;
